fix(query-generator): prevent name clash in subquery where-clause on self-reference
In the case of a limit on a self-reference with through-table, a subquery is generated with a subquery filter.The where-clause tries to reference the first instance (from) of the tabel but actually references the second instance (to) used in the join of the through-table as both tables are given the same default alias.
This fix checks for this case and gives the second table a randomized unique alias preventing confusion in the where-clause.